Add custom hooks, localstorage, useRef
Features:
Created the custom hook useFilteredExpenses, so that when you select a category, shows expenses only for this category.

Created a custom hook useLocalStorageState to get data from localstorage with using useEffect.

Use useRef to reference the title and focus on the input field with useEffect.

Use useRef and useEffect to create an animation of adding and deleting expenses in table